fixed: navigate_controller.cpp: In member function 'virtual bool KBEngine::NavigateController::update()': navigate_controller.cpp:111: warning: comparison between signed and unsigned integer expressions

+++ kbe/src/server/cellapp/navigate_controller.cpp
@@ -108,7 +108,7 @@ bool NavigateController::update()
 		if (!moveVertically_)
 			currpos.y = y;
 
-		if(destPosIdx_ == paths_.size() - 1)
+		if(destPosIdx_ == ((int)paths_.size()) - 1)
 			ret = false;
 		else
 			destPosIdx_++;
